scalefs: Enhance free-bit-vector management to support large filesystems
Today the free-bit-vector is a vector of free_bit structures, each of which
can consume upto 256 bytes (256 bytes in DEBUG mode, 48 bytes otherwise).
A filesystem of size 128MB has 32768 blocks (BSIZE = 4KB). So, we need exactly
8MB to represent these blocks in the free-bit-vector. But because operator
new stores the size of the allocation in the first word of the allocated
memory, we end up needing to allocate 16MB for this vector.

sv6 can support only a maximum contiguous allocation of 16 MB (with page size
being 4KB). Hence, the free-bit-vector can only support upto 32768 blocks
(128MB). (For reference, a related commit is : f243f688735 "mfs, boot: Improve
booting by allocating memory for free_bit_vector in bulk").

Thus, to support filesystems of larger size, we need to redesign free-bit
management. One simple solution is to maintain a vector of *pointers* to
free-bits rather than the entire free_bits themselves. Then, we would need
far lesser contiguous allocation for the vector, thereby allowing us to
support larger filesystems.

Interestingly, this has performance implications, and that too in contradicting
ways! The time taken to create a file, write 100 MB of data to it and then call
fsync on that file is as follows (this is part of the LFS largfile benchmark):

With DEBUG=1, without patch: 14.99 seconds
With DEBUG=1, with patch:    12.05 seconds (improvement)

With DEBUG=0, without patch:  3.14 seconds
With DEBUG=0, with patch:     5.28 seconds (degradation)

Thus, with this patch applied, we get a small improvement in DEBUG mode, but a
much more significant degradation with DEBUG mode off! Presumably this is
because of all the extra pointer dereferences that this patch generates,
because we now only store pointers to the free_bits in the vector, rather than
the free_bits themselves.

While these results might appear to not favor this patch, it turns out that
there is a much bigger bottleneck in the block allocation code, and this
patch paves the way to fix that in a subsequent patch.