scalefs, free-bit-vector: Perform block allocation and free in O(1) time
At present, block allocation is incredibly slow because we do a linear search
for free blocks in the free_bit_vector. To make matters much worse, every
block allocation scans the vector from the very beginning, even if we just
found out that the first 'n' blocks have been allocated!

Here is an example: Let's say we are writing 100 MB of data to a freshly
created file. So we'll need to allocate roughly 25000 blocks. Allocating the
block number 20001 for the file takes time proportional to traversing 20000
entries in the free-bit vector. For the very next block allocation (20002),
it traverses the first 20001 entries all over again! Thus, as the file grows,
block allocation gets noticeably slower.

We would like to allocate and free blocks in O(1) time. A simple technique to
do that is to maintain a vector as well as a linked-list representation of
the free-bits. The linked-list only contains the blocks that are actually
free. So allocating a block would amount to merely picking the first entry in
this freelist, thereby making it O(1). The vector representation is needed to
keep block-free fast, because freeing blocks depends on mapping the block-number
to the free_bit structure - the vector representation helps us do this in O(1)
time.

Not surprisingly, the performance improvement brought about by this patch is
tremendous, as shown in the LFS-largefile benchmark results below (for writing
and fsyncing a new 100 MB file):

Without patch:
=============

Test            Time(sec)       KB/sec
----            ---------       ------
seq_write       5.28            19381.23


With patch:
==========

Test            Time(sec)       KB/sec
----            ---------       ------
seq_write       0.39            256538.73  [ Improvement: 13x ]


There is a catch to this, though: since the linked-list cannot be updated
concurrently, both allocation and frees will get serialized at the freelist
lock. Thus, this patch makes block allocation and frees efficient (O(1) time)
but totally kills multi-core scalability. (As a result, the per free-bit
sleeplock is actually rendered sort of unnecessary by this patch). But note
that regular file writes remain scalable; this issue affects only sync/fsync
(because that's when we do block allocation).

However, we can fix the scalability issue (in a future patch) by partitioning
and maintaining per-cpu caches of free-blocks and allocating them locally on
a given CPU, refilling and flushing those caches as needed.