scalefs: Fix free_bit_vector initialization for large filesystems
When the filesystem is larger than 128MB, we need multiple blocks to hold the
free-block bitmap. The for-loop was initializing free_bit structures with
the bit index within a block, rather than using the cumulative bit index. So
larger filesystems would fail in weird ways during file read/write. Fix this.

Also, the bulk memory allocation for the free_bit_vector (using .reserve())
should be done outside the loop, to be truly effective for large filesystems.

And while at it, do some necessary code cleanup as well.