AHCI: Don't prematurely complete interrupts for cmds that have not been issued yet
handle_port_irq() examines the bits in preg->ci corresponding to the
cmdslot_dc[] entries that are populated, in order to check whether those
commands have finished successfully.

However, having a valid entry in cmdslot_dc[] does not guarantee that the
command was indeed issued, because there is a time-gap between populating the
entry and actually invoking ahci_port::issue(). Thus, we can end up prematurely
returning from the wait (synchronous or asynchronous) assuming that the
command has completed execution, which can lead to data corruption.

Fix this by adding a new variable called cmds_issued to note whenever a
command has been actually issued, so that the interrupt handler can
cross-verify with this before announcing the completion of the disk I/O.