AHCI: Write only the bit corresponding to the cmd slot of the cmd to be issued
The AHCI spec version 1.3.1 (section 5.5.1) says:

"Software shall set PxCI.CI(pFreeSlot) to indicate to the HBA that a
command is active. Software should only write new bits to set to ‘1’;
the previous register content of PxCI should not be re-written in the
register write".

Fix it by using a bitwise OR, instead of overwriting the entire value.
Moreover, the existing code is buggy in another sense: we detect that the
issued command has completed by checking that the corresponding bit in
PxCI has been cleared by the hardware. Thus, blindly overwriting the
variable would interfere with this mechanism.