AHCI: Fix memory corruption by allocating sufficient memory for 32 cmd slots
Originally, the ahci_port_page structure would fit within a page, so the
kalloc() of 1 page was sufficient. However, when support for multiple command
slots was added for the AHCI port, the structure bloated up but the memory
allocation was unchanged, and hence it would end up scribbling over
unallocated memory.

In retrospect, this looks like a very trivial issue, but its manifestations
were so bizzare that debugging from the other end proved to be a nightmare!

An initial rework of the ScaleFS transaction code to use asynchronous disk
I/O (using the 32 command slots) would crash in totally unrelated locations,
such as the page_info destructor's code to delete rmap_pte. Turning on
ALLOC_MEMSET revealed memory corruption of the "use-after-free" kind, but
printing out upto 2000 buddy allocations and frees immediately preceding the
crash didn't reveal anything useful, except that the corrupted address was
nearly always the same (non-random). Since page_info objects are always
allocated at the same fixed address, the crash would almost always be in the
page_info/rmap code.

Further investigation revealed that the probability of the crash was higher
whenever higher numbered command slots were being used. After some
pain-staking debugging to verify that the AHCI driver wasn't DMA'ing to wrong
addresses by cross-checking the command-slot initialization and the scatter-
gather list setup with the AHCI spec, eventually it dawned that the problem
was a trivial issue of insufficient memory allocation for the ahci_port_page
structure (which would mess up a lot of AHCI command data-structures for
the higher numbered command slots).

Fix this by allocating the right amount of memory and also put in an assert
statement to catch such bugs more easily in the future. Phew!