AHCI: Exhaust all cmdslot entries before sleeping on the condvar
When last_cmdslot == (hba->ncs)-2 and cmdslot[(hba->ncs)-1] is in use,
the driver goes to sleep on the condvar without checking the remaining
slots. Do a full pass in this case, to ensure that we exhaust all the
available slots, before sleeping on the condvar as a last resort.