fix subtle page_info memory corruption
page_info::onzero() was recently changed to call page_info::~page_info(),
but it did this in the wrong order:

  kfree(va());
  page_info::~page_info();

this way, as soon as kfree(va()) is called, another core can allocate the
same page.  it will receive the same exact memory for its page_info object,
since they are allocated using placement new.  this will collide with the
old instance of the object, which is still running page_info::~page_info().

the fix is to swap the order: first call the destructor, and then call
kfree(va()).  the latter is safe because the va() method computes the
page's address based on the address of the page_info object itself (this),
without referring to any page_info fields.