fix interesting fsync dependency bug
mfs_operation_delete claimed to be dependent on everything, but this
was a bad idea.  other operations (like mfs_operation_create) do not
claim to be dependent on everything, so it's possible to run into the
following scenario:

  file A is created (adds mfs_operation_create)
  file A is unlinked (adds mfs_operation_unlink)
  file A is deleted (adds mfs_operation_delete)
  file B is fsync'ed

at this point, B's fsync tries to apply A's mfs_operation_delete but
does not include mfs_operation_create.  now, mfs_operation_delete will
panic because it cannot find the on-disk inode number for A.

the fix is to make mfs_operation_delete depend on nothing; the only
reason to flush mfs_operation_delete is to free up some on-disk inodes.