bufcache: Use asynchronous disk I/O
It is easy to ensure that the buf is not evicted as long as the async I/O
is not complete (by incrementing the reference count in async_iowait_init()
and decrementing it in async_iowait()).

However, it is trickier to ensure that the buf is not permanently lodged
in the bufcache, in case a caller submits an async I/O but doesn't bother
to wait for its completion (thereby never invoking the corresponding dec()).
This risks jeopardizing buffer-cache eviction. So let's skip this part for
now as this needs some more thought. Therefore don't initiate buffer-cache
eviction concurrently with asynchronous disk I/O, for now.