ahci: Add workaround to use several command slots on ben, for async I/O
For some reason, we are unable to use all the 32 available command slots
per AHCI port on ben. However, it turns out that slots 0, 8, 16 and 24
are usable! Using any other command slot results in a timeout, since we
never get an interrupt for the commands issued via those command slots.

So, for now, just use these particular command slots on ben. With this,
we can issue 4 disk I/O commands in parallel, although eventually we
hope to be able to use all the 32 slots.