ahci: Write AHCI_CMD_FLAGS_WRITE to flags after calling fill_fis()
Otherwise, the update to the Write bit in the flags is not observed on ben,
and hence disk writes never complete (i.e., they don't generate completion
interrupts). This did not pose a problem in QEMU, however.

It appears that this order is important, because I also tried updating the
Write bit first, and then calling fill_fis() (but of course after changing
the update of 'flags' to a bitwise OR inside fill_fis()), and that didn't work.