mfs: Set file's dirty bit upon write, even if file size doesn't change
If the file size is not being changed, writei() in mfs takes care to set
the page's dirty bit when writing it, but fails to update the file's dirty flag.

This causes fsync to return silently, because it checks the file's dirty flag
first and if the file is dirty, only then it proceeds to examine every page's
dirty flag before writing it out.

Fix this by setting the dirty flag for the file upon write, even when the
file size is not being updated.