mfs: Fix the setting of the dirty bit for file pages
MFS sets the file page's dirty bit properly except when it is writing to an
already existing page. In this situation, it sets the dirty flag on a *copy*
of the page_state object, and hence the actual object in the pages_ radix array
is never updated. Thus, writes to file pages that already exist will not treat
the page as dirty.

Fix this by updating the dirty flag in the actual page_state object embedded
in the file's pages_ radix array.