benchmarks, LFS Smallfile: Invoke sync towards the end and time it
The time it takes to sync() everything at the end of creating and deleting
10,000 files is quite interesting because it can tell us a lot about how well
the filesystem handles committing a huge number of transactions.

On the other hand, note that create()s and delete()s cancel each other in
this benchmark. So a properly optimized filesystem would take advantage of
this fact to make sync() lightning quick!