scalefs: Implement group commit by batching many transactions in the journal
Today, every transaction is individually written out to the journal, then
the transaction is committed (which involves flushing all the writes to the
disk) and then the transaction is applied to the filesystem on the disk
(which involves another flush). Clear the journal, rinse and repeat.

The physical journal is sized in such a way that it can hold the largest
possible transaction (unlinking the largest possible file). However, if the
transactions to be committed are smaller, we can pack many of them in the
journal and then commit and flush them all out in bulk. This would improve
our disk write throughput.

This group commit scheme treats individual transactions (sub-transactions)
as parts of a top-level transaction, and commits the top-level transaction
when the journal can't accommodate any more sub-transactions.

It is to be noted that group commit changes the commit protocol and alters
how the physical journal looks. But it does not violate any atomicity
guarantees.

Before:
=======

  start block
  disk blocks of transaction 1
  commit block

<Intermediate, not implemented>:
===============================

  start block
  disk blocks of transaction 1
  commit block
  start block
  disk blocks of transaction 2
  commit block
  start block
  disk blocks of transaction 3
  commit block

After (with this patch):
========================

  start block
  disk blocks of transaction 1
  disk blocks of transaction 2
  disk blocks of transaction 3
  commit block


Running LFS smallfile benchmark (creating and deleting 10000 files) shows
dramatic improvements in the speed of the sync() stage, upto nearly 21x!

Without patch:
-------------
Test            Time(sec)       Files/sec
----            ---------       ---------
sync_files	597.27		16.74

With patch:
-----------

Test            Time(sec)       Files/sec
----            ---------       ---------
sync_files	28.31		353.19