scalefs: Encapsulate the writing of start and commit blocks into prolog/epilog
Every transaction begins with a start block and ends with a commit block. And
finalizing a transaction by writing the commit block takes several intricate
steps. There are quite a few places where we repeat these operations. So
encapsulate them into a prolog and an epilog, to enhance code reuse and
readability.