scalefs: Try to absorb deletes with corresponding create/link/unlink/rename
If the transaction log has create for an mnode and the corresponding delete
as well, we can absorb them both, because they cancel each other out. Same
goes for the other transaction types as well. But note that absorption is
safe only when we have *all* the transactions of the mnode, right from the
create until the delete. Else, we will have to execute all the transactions.