benchmarks, LFS Smallfile: Sleep for a few seconds before calling sync()
For the create-delete transaction absorption optimization to work, we need
refcache to delete the mnode after the file is unlinked, because that's
when the delete transaction is added to the log. Refcache might take a
while to kick in. So give a small breather between unlink and sync to give
refcache enough time to do its job.