benchmarks, LFS Smallfile: Implement concurrent benchmark runs using pthreads
Provide a -c (cpus) parameter to indicate how many CPUs to run the benchmark
on. The benchmark will then be run on those many CPUs (specifically, on cpus
numbered 0 through <c> - 1, using CPU affinity to bind the threads).

The files created by the different threads are distinguished by using the
cpu-number as a part of the filename itself (in order to avoid interference).

Also, print out the total time taken by each thread (excluding the time spent
in sleep() within the benchmark, if any).