scalefs: Distinguish sync and fsync in order to optimize flushing the journal
During fsync(), we need to flush the journal to disk immediately, because
it could be a single isolated fsync() that we are processing. However, in
the case of sync() (particularly sync_dirty_files()), we know that we will
be synchronizing a large number of files, and it would be more efficient to
flush the transactions from the journal to the disk in one-shot (using group
commit), after processing all the files.

Add support to distinguish these 2 cases appropriately, so that we can
optimize flushing the journal to disk during sync().