scalefs: Don't add the transaction to the transaction_log during fsync
add_fsync_to_journal() can be called either in the fsync() or in the sync()
path. When it is called in the fsync() path, there is no need to add the
transaction to the transaction_log, because we will commit, apply and delete
the transaction immediately.

This was actually the original behavior, but commit ae1629b changed it when
it introduced add_to_journal_locked() inside add_fsync_to_journal(). But
unfortunately, that commit failed to remove the transaction from the
transaction_log before returning from add_fsync_to_journal(). Hence that ends
up causing trouble to a sync following an fsync, as shown below, because it
tries to delete a non-existent transaction from the transaction_log:

    $ echo a  > b.txt
    $ fsync b.txt
    $ sync
    kernel trap 13 err 0x0 cpu 0 cs 16 ds 24 ss 24
      rip ffffffffc0176219 rsp fffff0000015cd88 rbp fffff0000015cda8
      cr2 0000000001000000 cr3 000000000f386000 cr4 00000000000006b0
      rdi ffffff000f4ffc80 rsi 0000000000000000 rdx ffffff0010f7bc08
      rcx 0000000000000001 r8  ffffffffc0c8aac8 r9  0000000000000080
      rax ffffff0012747240 rbx 0303030303030303 r10 ffffff0000000000
      r11 7ffffffffffff000 r12 ffffff000f4ffc80 r13 ffffff000f4ffc80
      r14 0000000000000000 r15 ffffff0012747440 rflags 0000000000000206
      proc: name sync pid 35 kstack 0xfffff00000155000
      ffffffffc017484f
      ffffffffc01759f9
      ffffffffc01423ef
      ffffffffc0141b38
      ffffffffc015affc
      ffffffffc015c5e3
      00000000004003d1
      0000000000400b12
      0000000000400372
      ffffffffffffffff

(The IP points to "delete (*it);" inside flush_journal_locked(), which tries
to delete a transaction found in the transaction_log).

Fix this by never adding the transaction to the transaction_log in the fsync
path, because it is unnecessary to begin with.