scalefs: Use PGSIZE as the size argument to sync_file_page(), to speed up writes
sync_file_page() invokes writei() to write out the contents to blocks on disk.
writei() can skip reading the disk block if it notices that we are about to
completely overwrite the disk-block's existing contents, i.e., if the size of
the write is equal to PGSIZE (which is also the same as BSIZE).

Skipping the synchronous disk read can be particularly helpful in speeding up
sync/fsync, because that way we can exploit the full power of asynchronous disk
writes, by avoiding Read-Modify-Write.

sync_file() invokes sync_file_page() with nbytes as the size argument, which
can be less than PGSIZE. (In particular, in the LFS Smallfile benchmark, nbytes
is 1 KB). However, it is to be noted that we always allocate new file pages
using zalloc() [ see writei() in kernel/mfs.cc ] and our only truncation scheme
is O_TRUNC, which completely truncates the file to zero length.

Thus, even if we want to sync only a part of a page to the disk, we can actually
write out the entire page because the rest of the bytes will be zero anyway (and
hence harmless). So always supply PGSIZE as the size argument to sync_file_page().
This will help us avoid Read-Modify-Write in writei(), thereby greatly speeding
up sync/fsync, particularly in the LFS Smallfile benchmark.

Experiments show that the sync stage (after create) in the LFS Smallfile
benchmark takes about 3-5 minutes before this patch, and only a few seconds
after this patch, when running on a real disk.