fs, memide: Support a larger filesystem in memory
Increase the filesystem size from 128 MB to 1 GB (so that we can run bigger
experiments), and add the appropriate code to also support this in a RAM-disk
fashion in memory.

The kernel image (which also includes the filesystem disk image) is compressed
during build and will get decompressed upon boot, to an area of memory large
enough to hold the decompressed image.

However, regular kmalloc() doesn't support very large contiguous allocations
of the size of a gigabyte. So, we need to use the early memory allocator
(early_kalloc()) to achieve this. To do so, we need to also ensure that we
allocate this memory early enough (before calling initkalloc()), so that we can
actually make use of the early memory allocator for this purpose. (Otherwise,
its internal checks will fail and it will fallback to the regular memory
allocator, which won't be able to satisfy our huge allocation request).