chainhash: Add a method to remove a key-value pair, by only specifying the key
Chainhash already has a remove() method, but it expects both the key and the
value to be specified by the caller. This works great for updating directory
entries, but in general, we also need to support a remove() method that only
takes the key as the argument. We will be using this method shortly.