mm, buddy: Return early if high order allocation is not satisfiable
If a task requests memory of order N, and the buddy allocator doesn't have
any free pages of order >= N, then there is no point in looking up free lists
of higher orders (from N + 1 through MAX_ORDER), because we are certain to
return nullptr eventually. So speed up the buddy allocator for this scenario
by maintaining a variable that tracks the highest order allocation that the
buddy allocator can satisfy successfully.

While running the LFS-Smallfile benchmark, we allocate nearly 30000
transactions, all of which perform an order-6 allocation for their embedded
hash-tables. On ben, the memory on NUMA Node 0 is fragmented (due to holes)
and hence sv6 carves out 3 different buddy allocators, roughly of size
200 MB, 100 KB and 30 GB respectively. As expected, during the benchmark
we quickly exhaust memory from the first 2 allocators, and start using the
third one. However, from that point onwards, every allocation request checks
the first 2 allocators and only then requests memory from the 3rd allocator
(due to the preference list). This patch will make it faster to fallback to the
3rd allocator, because we won't have to recursively look up free lists from
order 6 to MAX_ORDER (12) in each of the first 2 allocators.