scalefs: Allocate the hash-table within a transaction on-demand
Allocating a transaction object is expensive because it involves dynamically
allocating the hash-table embedded within it (which needs an order 6 allocation).
We would like to avoid this whenever possible, by using smarter ways to
deduplicate the transaction diskblocks, especially by utilizing group commit.
As a first step towards that, allocate the hash-table only when necessary
(i.e., when using one of the add_unique_* methods).

Next, avoid using the hash-table by changing the deduplication strategy: sort
the diskblocks in the order of (blocknum, timestamp) and discard everything but
the latest versions of the disk block contents based on the timestamp. This is
actually a revival of an old piece of code: we were using this deduplication
strategy sometime back, until commit dc0bf6209 (scalefs: Remove the now-obsolete
disk-block sorting in __flush_transaction()) replaced it with a better scheme
(which used the hash-table). But since the hash-table has other issues (memory
allocations being slow) which affect its overall performance, this patch
effectively reverts to the earlier strategy (sorting), at least temporarily.