scalefs: Fix the committing of transactions when they don't overflow the journal
When the number of transactions to be committed are small, there will be plenty
of space left in the journal and hence we will quit the for-loop without ever
entering the 'else' branch in flush_journal_locked(). The code immediately
following the for-loop doesn't handle this scenario well, because it writes out
the epilog, omitting the transaction diskblocks. Fix it by committing these
transactions properly.