fs: Comment out obsolete functions related to the removal of directory entries
The functions dir_remove_entry() and dir_remove_entries() are no longer used.
We are trying to rewrite directory handling, and these functions get in the way
of compilation. So comment them out for now. We'll delete them later once we
are sure that the rewrite is better than the existing code.