disk: Flash the sv6 filesystem image to disk during boot
This makes it easier to try out patches that make changes to the fs image
(such as modifications to the userspace programs etc) without having to 'dd'
the image to disk from Linux. It also helps get past an sv6 filesystem that
has been corrupted due to a crash, without having to rely on Linux.