boot, zlib: Decompress the fs image in smaller chunks
Currently we decompress the filesystem image all at once, which requires that
the output buffer is pre-allocated and ready to go. It would be far more useful
to be able to decompress and fill the output one chunk at a time, especially when
the output buffer is fragmented (such as having multiple MEMIDE disks, or a
very large filesystem image that exceeds the size of any contiguous memory
allocation on any NUMA node).

So modify zlib_decompress() to fill the output buffer in smaller chunks.