boot, zlib: Copy output to destination using caller-supplied callback
As we begin to deal with multiple disks, we would like to lay out the
decompressed filesystem image in various (often complex) ways across
the disks (such as linear, RAID-0 striping etc). So modify zlib_decompress()
to accept a callback function, which it invokes to copy the decompressed
output chunk by chunk. We can then implement any layout we like inside
the callback function.