scalefs: Always call create_file/dir_if_new() with link_in_parent set to true
Calling create_file/dir_if_new() followed by dirlink() is equivalent to calling
the former with the link_in_parent flag set to true. So get rid of those
unnecessary calls to dirlink() in create_directory_entry() [ but take care to
avoid self-deadlock due to ilock() on the parent directory ].

This is a patch in preparation to remove the link_in_parent flag altogether.