mm: Increase the maximum size of contiguous allocation
Once in a while, when running the LFS benchmarks under heavy stress, the kernel
panics because it can't allocate a large enough chunk of contiguous memory.
Increasing MAX_ORDER in the buddy allocator to 14 fixes this problem.

However, this is only a temporary solution. The real issue is that the
prune_trans_log hashtable (used for deduplication) can grow very large depending
on the number of transactions that get queued up, thus needing huge contiguous
allocations. But luckily we will soon be getting rid of that hashtable, with some
upcoming changes.

Further, there are a few vectors in the kernel that can also grow unbounded (thus
demanding large contiguous allocations), but some of them don't really need to
be vectors; they will work just fine as linked-lists too. So once all these are
fixed, remember to revert back MAX_ORDER to 12, since the worst-case allocation
time of the buddy allocator increases with this number.