scalefs, fsync: Scalable tracking of metadata operations using per-file oplog
Today, we have a single metadata_log data-structure replicated on each CPU,
which is used to log all the metadata operations that occur on that CPU. This
scales very well when logging the metadata operations, but doesn't scale as
much when trying to merge the logged operations together during an fsync().

That is, when running fsync()s on different files in parallel, all the fsync()
calls get serialized in order to perform the merge, and furthermore, each
fsync()'s dependency tracking algorithm also ends up running serially. Thus
the current design of process_metadata_log(u64, u64, bool) is a major bottleneck
to fsync()'s scalability.

To make concurrent fsync()s scale, the first step is to allow the fsync() calls
to track their dependent metadata operations in parallel, which in turn demands
a redesign of the metadata logging mechanism.

Instead of having a single oplog maintained metadata_log data-structure per-cpu,
we will have a metadata_log for every mnode in the system, and these numerous
metadata_log data-structures will be arranged in a hash-table with mnode number
as the key. So the new picture looks something like this:

Global hash-table, mapping mnode numbers to their respective logging data-
structures:

   mnode 1000 |-> metadata_log
   mnode 1050 |-> metadata_log
   mnode 2000 |-> metadata_log
   mnode 2050 |-> metadata_log
   mnode 3000 |-> metadata_log

Each such metadata_log data-structure is replicated on every CPU, since it is
maintained by oplog. Thus we move from having per-cpu metadata_log, to per-file,
per-cpu metadata_log.

With this infrastructure in place, concurrent fsync()s on different files will
access different metadata_logs (based on the mnode numbers), and thus the merging
of metadata operations as well as the dependency tracking can proceed in parallel.
(As an additional advantage, the amount of metadata operations merged during an
fsync() [and hence the work involved] becomes proportional to the fsync'ed
file's dependencies. It won't have to blindly merge all operations any more).

This patch implements the initial bits of the per-file, per-cpu metadata_log.
We allocate the logging structure as soon as a new mnode number is allocated
in the system.