scalefs: Modify metadata logging to accommodate the new per-file oplog
Since metadata_log is now maintained separately for every mnode in the system,
we need to modify metadata_op_start(), metadata_op_end() and
add_to_metadata_log() to also specify the corresponding mnode number, so
that we can log the operations in the appropriate logging data-structures.

Note that we also need to adjust the failure handling paths in sys_rename()
and create(), to ensure that an _op_start() is always paired (followed) by
an _op_end() in those failure paths.

From here onwards, we refer to the per-file, per-cpu logging data-structure
as mfs_log, instead of metadata_log, to avoid confusion.