scalefs: Update the dependency checks for file fsync()
Now that the mfs_log is organized by mnode, the dependency checks become a lot
more simple, because at any given time, the set of metadata operations on
which the dependency checks are invoked, all belong to a single mnode.

Also, take this opportunity to streamline the checks for unlink() and rename(),
such that they only add the bare minimum mnodes to the dependency list.