scalefs: Log all delete operations in a common, dedicated (per-cpu) oplog instance
We are about to modify the logging of MFS operations in the oplog, such that
every metadata operation (create, link, unlink etc) is associated with its
parent directory mnode, rather than with its file (or child directory) mnode.
All operations except 'delete' have an associated directory, which leaves
'delete' as the odd man out.

So define a dedicated (per-cpu) oplog instance specifically meant to log all
'delete' operations.

Note that this won't hamper our efforts to perform dependency tracking between
metadata operations during fsync, because it turns out that, under the newly
defined semantics for fsync, 'delete' operations cannot be flushed out during
an fsync! Here is why:

The delete operation for a file manifests itself after all links to the file's
inode have been unlinked and all the file descriptors referring to that inode
have been closed. But since fsync() takes a file descriptor as its argument,
there is no way to fsync() a file once it is ready for deletion. (And because
of the new semantics of fsync, fsync()ing the file's parent directory will
not flush out the 'delete' operation corresponding to that file).

Thus, it follows that file deletion can be synchronized to disk only during a
sync(). So we won't need to perform parallel dependency tracking of 'delete's
during fsync(), and hence it is acceptable to log all 'delete's under a
common (per-cpu) mfs_log.

To achieve this, define MFS_DELETE_MNUM to be the special key which helps us
lookup the per-cpu mfs_log dedicated for deletes. Since zero is an invalid
mnode number, we can set MFS_DELETE_MNUM to zero, thus ensuring that it won't
clash with anything else.