scalefs: Associate 'create' metadata operations with the parent directory
At the moment, 'create' operations are logged under the mfs_log corresponding
to the new file's (or the new child directory's) mnode. Change the logging
scheme to log the operations under the mfs_log corresponding to the parent
directory's mnode instead. (That is, switch the anchor for the metadata
operation, from the child to the parent).