scalefs: Associate 'rename' metadata operations with the source directory
At the moment, 'rename' operations are logged under the mfs_log corresponding
to the file's (or the child directory's) mnode. Change the logging scheme to
log the operations under the mfs_log corresponding to the source directory's
mnode instead. (That is, switch the anchor for the metadata operation, from
the child to the source directory).