scalefs: Dont truncate the old file within create_directory_entry()
If create_directory_entry() finds that the link already exists but it points
to a different inode, it proceeds to unlink it and also truncates the file
if its link count drops to zero.

However, this is incorrect because userspace might still have open file
descriptors for that file. So call upon unlink_old_inode() to perform the
unlink properly (it handles truncation correctly).