scalefs: Add a helper to allocate an inode for a given mnode
Define a helper to allocate an inode corresponding to an mnode, and invoke it
from create_file_dir_if_new(). Note that we don't need to have an open
transaction to allocate inodes; So we can invoke this helper at a much earlier
stage, if necessary.

While at it, invoke iupdate() with only the readlock held over the inode, to
allow concurrent readers.