scalefs: Repurpose 'create', and then connect 'create' and 'link' operations
Creating a new file or directory involves 2 important steps:

1. Allocating a file or directory inode (and if it is a directory, adding the
   ".." link in it).

2. Adding a link to the newly created file or directory in its parent directory.

Let us repurpose the MFS 'create' metadata operation such that it only performs
step 1, and let 'link' metadata operation handle step 2, since it is already
designed to do so.

This work distribution will suit the new semantics of fsync better (which
demands that the file and its parent directory be fsync()'ed individually, to
synchronize the file contents to disk and make sure that the file is reachable
via path-names on the disk, respectively).

To achieve this, we will also have to alter the logging scheme a bit, so that
every file or directory creation is logged as 2 metadata operations - a 'create'
and a 'link', with the exact same timestamp. Note that the 'create' operation
will be logged with the child and the 'link' operation will be logged with the
parent.