scalefs: Allocate parent directory's inode (if needed) when creating sub-directory
The 'create' metadata operation has to allocate a new inode for the file or
directory being created, and also add the mandatory ".." link when creating a
directory.

However, to add that ".." link, we need to know the inode number of its parent
directory. So if the parent directory doesn't have a corresponding inode yet,
allocate one on-demand.

In this particular scenario, we will refrain from adding the ".." link in the
parent directory in turn, during its inode creation, because of the following
reasons:

1. If we plan to add the ".." link in the parent as well, then we will have to
   recursively go up the filesystem hierarchy and allocate inodes for every
   directory up the chain, until we hit a directory that already has an inode.
   This gets complicated pretty quickly and it doesn't sit very well with fsync's
   goal of flushing out a minimal set of relevant changes to disk.

2. Luckily, it turns out that adding the ".." link in the parent is not
   necessary at this point, because until the parent directory is fsync()'ed
   separately, the child directory will remain unreachable via path-names on
   the disk. So even if we crash at this point (after fsync()ing the child
   directory), we won't be exposing any inconsistent filesystem state to
   userspace (such as directories that don't have the mandatory ".." links),
   because that "inconsistent state" (i.e., the parent directory) will also be
   unreachable.

However, we do make sure that if the parent directory is fsync()'ed, an inode
will be allocated for it (if it doesn't already have one), and its ".." link
will be properly initialized. Further, if the parent directory is fsync()'ed
first, then while adding the link to the child directory in the parent, the
child's inode will be allocated and its ".." link initialized, since we will
already know the parent directory's inode number.