scalefs: Consolidate all implementations of get_tsc()
There are several implementations of get_tsc() in the kernel, and also some
variations, such as get_timestamp(). Consolidate all of them into a common
implementation.

(hpet.hh is not the ideal place for it to reside, but that's better than
sticking it in the all-too-generic kernel.hh or among the low-level code in
say cpuid.hh or amd64.h).